Add function - check session to login info

+++ Servlet_CRUD/src/member/ServletDelete.java
@@ -13,8 +13,10 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import dao.MemberDao;
+import vo.Member;
 
 
 @WebServlet("/Delete")
@@ -22,15 +24,21 @@ public class ServletDelete extends HttpServlet {
 	private static final long serialVersionUID = 1L;
    
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		ServletContext sc = this.getServletContext();
-		Connection conn = (Connection)  sc.getAttribute("conn");
-		MemberDao dao = new MemberDao();
-		dao.setConnection(conn);
-		try {
-			dao.deleteMember(Integer.parseInt(request.getParameter("mno")));
-			response.sendRedirect("List");
-		}catch(Exception e) {
-			throw new ServletException(e);
+		HttpSession session = request.getSession();
+		Member member =(Member)session.getAttribute("member");
+		if(member != null) {
+			ServletContext sc = this.getServletContext();
+			Connection conn = (Connection)  sc.getAttribute("conn");
+			MemberDao dao = new MemberDao();
+			dao.setConnection(conn);
+			try {
+				dao.deleteMember(Integer.parseInt(request.getParameter("mno")));
+				response.sendRedirect("List");
+			}catch(Exception e) {
+				throw new ServletException(e);
+			}
+		}else {
+			response.sendRedirect("Login");
 		}
 	}
 }
